clients/python/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
astroid==3.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
dill==0.3.8
distro==1.9.0
frozenlist==1.4.1
googleapis-common-protos==1.62.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
isort==5.13.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.22
langsmith==0.0.87
marshmallow==3.20.2
mccabe==0.7.0
multidict==6.0.5
mypy==1.8.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.12.0
packaging==23.2
platformdirs==4.2.0
protobuf==4.25.2
pydantic==2.6.1
pydantic_core==2.16.2
pylint==3.0.3
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.1.0
sniffio==1.3.0
SQLAlchemy==2.0.27
tenacity==8.2.3
tomlkit==0.12.3
tqdm==4.66.2
types-protobuf==4.24.0.20240129
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
